Commit transaction only when changing database
Instead of committing on each query, just wait for another a switch to another
database to commit. We need to commit on change to support DROP OWNED on all
databases before DROP ROLE.

This changes means that an error will loose previous queries in transaction.
This is not a problem as long as configuration is reproduceable. And it should.